% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{A variable indicating the number of desired iterations.}

\item{n}{Number of desired sample size per iteration}

\item{p}{The probability of sucess for each sample size}
}
\value{
A barplot of the binomial simulation
}
\description{
Simulates binomial outcomes for a specified number of iterations and sample size with a given probability of success. It generates a bar graph displaying the proportions of successes across all iterations.
}
\examples{
\dontrun{mybin(iter=1000,n=18, p=0.3)}
}
