% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Bootstrap Function}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{The number of iterations}

\item{x}{The data vector provided}

\item{fun}{The function to be applied to each bootstrap}

\item{alpha}{The significance level for the confidence interval}

\item{...}{Optional additional arguments/parameters}
}
\value{
A list of ci, x, and fun along with a plot of the bootstrap
}
\description{
Bootstrap Function
}
\examples{
\dontrun{myboot(x = sam, fun = "mean", alpha = 0.05, xlab = "mean", col = "palegoldenrod")}
}
