% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Function 2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{The number of iterations}

\item{x}{The data vector provided}

\item{fun}{The function to be applied to each bootstrap}

\item{alpha}{The significance level for the confidence interval}

\item{cx}{The size of the text label on the plots}

\item{...}{Optional additional parameters}
}
\value{
An invisible list of ci, fun, and x along with a histogram plot
}
\description{
Performs a bootstrap simulation to calculate confidence intervals for a use-defined statistic, that is fun, applied to a dataset. This is the better looking function.
}
\examples{
 \dontrun{myboot2(x = sam, fun = "mean", alpha = 0.05, xlab = "mean", col = "palegoldenrod")}
}
