% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence Interval Function}
\usage{
myci(x, alpha = 0.05)
}
\arguments{
\item{x}{Random data that you want to create a confidence interval for}

\item{alpha}{The alpha value of the function}
}
\value{
A confidence interval with the given data and alpha value
}
\description{
Calculates a confidence interval for the mean of the dataset "x" based on a specified significance level.
}
\examples{
\dontrun{myci(x, alpha = 0.05)}
}
