% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Central Limit Theorem - Binomial}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{Defines how many trials are performed in each binomial sample.}

\item{iter}{The number of samples that the function will generate.}

\item{p}{The probability of success in each trial (default value is 0.5)}

\item{...}{Optional additional arguments that is passed to hist()}
}
\value{
Outputs a histogram plot showing the distribution with a theoretical normal curve
}
\description{
Simulates the Central Limit Theorem by generating multiple binomial samples, calculating the means, and plotting a histogram of these means.
}
\examples{
\dontrun{mycltb(n = 10, iter = 10000, p = 0.3)}
}
