% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlike.R
\name{mymaxlike}
\alias{mymaxlike}
\title{Max Likelihood}
\usage{
mymaxlike(lfun, x, param, ...)
}
\arguments{
\item{lfun}{The function to focus on}

\item{x}{The data being passed to the function}

\item{param}{The range (vector) of parameter values for the likelihood}

\item{...}{Optional parameters}
}
\value{
The max likelihood for the provided parameters in a list and plots the graph
}
\description{
Calculates the maximum likelihood estimate for the given likelihood function and dataset.
}
\examples{
\dontrun{mymaxlike(x = c(4, 6, 7, 6, 5), param = seq(0, 20, length = 1000), lfun = logpoiss, xlab = expression(lambda), main = "Poisson", cex.main = 2)}
}
