% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysquare.R
\name{mysquare}
\alias{mysquare}
\title{The Square Function}
\usage{
mysquare(x)
}
\arguments{
\item{x}{a vector of quantitiative data}
}
\value{
a vector of qualitative data
}
\description{
The Square Function
}
\examples{
mysquare(x = 1:10)
}
