% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets Function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats in the flight}

\item{gamma}{The probability a plane will be truly overbooked}

\item{p}{The probability of a "show", meaning someone shows up}
}
\value{
A list containing N, gamma, p, nc (normal approximation), and nd (discrete distribution), where nc and nd is the calculation of the number of tickets to be sold. Creates two plots for the normal approximation and discrete distribution.
}
\description{
Determines the optimal number of tickets to sell for a flight, given the total number of seats and overbooking probability, and the probability of a passenger showing up.
}
\examples{
\dontrun{ntickets(400, 0.02, 0.95)}

}
