% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterHist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter Histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{X variable as a vector}

\item{y}{Y variable as a vector}

\item{xlab}{X label}

\item{ylab}{Y label}
}
\value{
A sophisticated plot
}
\description{
Creates a plot that combines a scatter plot with histograms for both X and Y variables.
}
\examples{
\dontrun{with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))}
}
