% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The standard deviation of the normal distribution}

\item{a}{The upper bound of the shaded region}
}
\value{
A list that has the mean, standard deviation, probability, and a graph of the shaded curve
}
\description{
Visualizes the normal distribution for a given mean and standard deviation and calculates the probability for a specified upper bound.
}
\examples{
\dontrun{myncurve(mu=10,sigma=5, a=6)}
}
